examples/sol-compiler-python/requirements.txt
lark-parser==0.11.1
